Player and Obstacle Classes
Created the player and obstacles. Then I set up the obstacles to be
randomly placed and then come at the player, once they leave the screen
they go back to the top and loop through again. Very basic, needs a lot
of tuning as to the spacing of the obstacles.  The obstacles are
restricted to 5 distinct tracks. Right now the player moves freely but
we will probably want hitting left and right to switch over by one track
in either direction.